requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.38.0
Pillow>=9.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
tqdm>=4.65.0
